+++ src/app/(Product)/wedding/invite/[id]/_components/InviteGuestBook/components/BookComments/hooks/index.ts
@@ -2,3 +2,4 @@ export { default as useDeleteComment } from './useDeleteComment/useDeleteComment
 export { default as useAdminDeleteComment } from './useAdminDeleteComment/useAdminDeleteComment';
 export { default as useCommentId } from './useCommentId/useCommentId';
 export { default as useCheckLogin } from './useCheckLogin/useCheckLogin';
+export { default as useCommentPages } from './useCommentPages/useCommentPages';
