+++ src/app/(Product)/wedding/invite/[id]/_components/InviteGuestBook/components/BookComments/hooks/useCommentPages/useCommentPages.type.ts
@@ -1,0 +1,7 @@
+import { GuestBook } from '@/types/response';
+
+export interface UseCommentPagesProps {
+  isLoading: boolean;
+  data: GuestBook | undefined;
+  page: number;
+}
